[angel] add gmail config

+++ app/Http/Controllers/ArticleController.php
@@ -106,7 +106,8 @@ class ArticleController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Article $article)
-    {
+    {   
+        $this->authorize('delete', $article);
         $article->delete();
         return redirect()->route('articles.index')
                         ->with('success', 'Article deleted successfully');
